Split GT viz into annotation
