+++ setup.cfg
@@ -50,8 +50,8 @@ dev =
   mrcfile
   starfile>=0.4.12
   tensorboard>=2.12.3
-  magicgui>=0.7.2
-  napari>=0.4.17
+  # magicgui>=0.7.2
+  # napari>=0.4.17
   # ffmpeg>=4.2.2
 
 [options.entry_points]
